# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Renta de Apartamentos - API",
        "description": "Documentación de la API del sistema de rentas de apartamentos. Permite gestionar usuarios, apartamentos, quejas, contratos y más.",
        "contact": {
            "email": "tucorreo@ejemplo.com"
        },
        "license": {
            "name": "Todos los derechos reservados - Universidad Piloto de Colombia"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/rol": {
            "get": {
                "tags": [
                    "Rol"
                ],
                "summary": "Listar todos los roles",
                "operationId": "feb22471238d3238429bce6ee92ea94b",
                "responses": {
                    "200": {
                        "description": "Lista de roles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Rol"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Rol"
                ],
                "summary": "Crear un nuevo rol",
                "operationId": "840af6b233185ed860b79592f49f9cf4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Rol"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rol creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Rol"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/rol/{id}": {
            "get": {
                "tags": [
                    "Rol"
                ],
                "summary": "Mostrar un rol específico",
                "operationId": "66ac972107932749b152d60f66b206d1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del rol",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rol encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Rol"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Rol no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Rol"
                ],
                "summary": "Actualizar un rol",
                "operationId": "b807d340545f43c929f1422bdc8fc690",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del rol",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Rol"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rol actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Rol"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Rol"
                ],
                "summary": "Eliminar un rol",
                "operationId": "8125fad829dc94c51a6530a4af4f9bee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del rol",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rol eliminado correctamente"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Rol": {
                "title": "Rol",
                "description": "Esquema del modelo Rol",
                "required": [
                    "nombre"
                ],
                "properties": {
                    "id": {
                        "description": "ID autoincremental del rol",
                        "type": "integer",
                        "readOnly": true
                    },
                    "nombre": {
                        "description": "Nombre del rol",
                        "type": "string",
                        "maxLength": 255
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}